fixes to Rooms and Bookings

+++ src/main/java/enset/bdcc/pi/backend/ApiClass/ApiBooking.java
@@ -6,11 +6,11 @@ import java.sql.Date;
 
 @Data
 public class ApiBooking {
-    private int id;
+    private Long id;
     private Date date;
     private String start;
     private String end;
     private String title;
-    private int roomId;
+    private Long roomId;
     private Long userId;
 }
